Complete version one of our security module.
There are five accessible functions:
requestUserKey puts the user into a queue of users for which keys need to be generated.
generateUserKey generates an OpenPGP-compliant keypair for the user.
encryptString encrypts a string, returning a promise that resolves to the ciphertext.
decryptString decrypts a string, returning a promise that resolves to the plaintext.
hashString returns a SHA-512 hash of the input.